# Invoice-System-Pro/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandro Industrial Invoice System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type InvoiceStatus = (typeof InvoiceStatus)[keyof typeof InvoiceStatus];

export const InvoiceStatus = {
  pending: "pending",
  paid: "paid",
  cancelled: "cancelled",
} as const;

export type StaffRole = (typeof StaffRole)[keyof typeof StaffRole];

export const StaffRole = {
  measurer: "measurer",
  quoter: "quoter",
  both: "both",
} as const;

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  seller: "seller",
  readonly: "readonly",
} as const;

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod];

export const PaymentMethod = {
  "50%": "50%",
  "100%": "100%",
  contra_entrega: "contra_entrega",
} as const;

export interface LoginRequest {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
}

export interface LoginResponse {
  token: string;
  user: User;
}

export interface Client {
  id: number;
  code?: string;
  name: string;
  address?: string;
  city?: string;
  province?: string;
  country?: string;
  phone?: string;
  mobile?: string;
  createdAt?: string;
}

export interface CreateClientRequest {
  code?: string;
  name: string;
  address?: string;
  city?: string;
  province?: string;
  country?: string;
  phone?: string;
  mobile?: string;
}

export interface Product {
  id: number;
  code?: string;
  name: string;
  description?: string;
  unitPrice: number;
  unit?: string;
  category?: string;
  stock?: number;
  active?: boolean;
}

export interface CreateProductRequest {
  code?: string;
  name: string;
  description?: string;
  unitPrice: number;
  unit?: string;
  category?: string;
  stock?: number;
  active?: boolean;
}

export interface StaffMember {
  id: number;
  name: string;
  role: StaffRole;
  active?: boolean;
}

export interface CreateStaffRequest {
  name: string;
  role: StaffRole;
  active?: boolean;
}

export interface InvoiceItem {
  id?: number;
  productId?: number;
  description: string;
  quantity: number;
  unitPrice: number;
  total: number;
}

export interface Invoice {
  id: number;
  invoiceNumber: string;
  createdAt?: string;
  updatedAt?: string;
  clientId?: number;
  client?: Client;
  measuredById?: number;
  measuredBy?: StaffMember;
  quotedById?: number;
  quotedBy?: StaffMember;
  budgetDate?: string;
  deliveryDate?: string;
  paymentMethod?: PaymentMethod;
  downPayment?: number;
  notes?: string;
  status: InvoiceStatus;
  items?: InvoiceItem[];
  total: number;
}

export type CreateInvoiceRequestItemsItem = {
  productId?: number;
  description: string;
  quantity: number;
  unitPrice: number;
  total: number;
};

export interface CreateInvoiceRequest {
  clientId?: number;
  measuredById?: number;
  quotedById?: number;
  budgetDate?: string;
  deliveryDate?: string;
  paymentMethod?: PaymentMethod;
  downPayment?: number;
  notes?: string;
  status?: InvoiceStatus;
  items: CreateInvoiceRequestItemsItem[];
}

export interface PaginatedInvoices {
  data: Invoice[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface PaginatedClients {
  data: Client[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface PaginatedProducts {
  data: Product[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}

export interface MonthlyRevenue {
  month: string;
  total: number;
  count: number;
}

export interface DashboardMetrics {
  monthInvoices: number;
  monthRevenue: number;
  pendingInvoices: number;
  paidInvoices: number;
  monthlyRevenue: MonthlyRevenue[];
  recentInvoices: Invoice[];
}

export type ListInvoicesParams = {
  page?: number;
  limit?: number;
  search?: string;
  status?: InvoiceStatus;
  dateFrom?: string;
  dateTo?: string;
};

export type UpdateInvoiceStatusBody = {
  status: InvoiceStatus;
};

export type ListClientsParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type ListProductsParams = {
  page?: number;
  limit?: number;
  search?: string;
  category?: string;
  active?: boolean;
};