# Invoice-System-Pro/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Sandro Industrial Invoice System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      summary: List invoices with pagination
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/InvoiceStatus"
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Paginated invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedInvoices"

    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new invoice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"

    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Soft-delete invoice
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /invoices/{id}/status:
    patch:
      operationId: updateInvoiceStatus
      tags: [invoices]
      summary: Update invoice status
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  $ref: "#/components/schemas/InvoiceStatus"
              required: [status]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated clients
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedClients"

    post:
      operationId: createClient
      tags: [clients]
      summary: Create client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"

    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /clients/{id}/invoices:
    get:
      operationId: getClientInvoices
      tags: [clients]
      summary: Get invoices for a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    InvoiceStatus:
      type: string
      enum: [pending, paid, cancelled]
    StaffRole:
      type: string
      enum: [measurer, quoter, both]
    UserRole:
      type: string
      enum: [admin, seller, readonly]
    PaymentMethod:
      type: string
      enum: ["50%", "100%", "contra_entrega"]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    LoginResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
      required: [id, email, name, role]
    Client:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
        province:
          type: string
        country:
          type: string
        phone:
          type: string
        mobile:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name]
    CreateClientRequest:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        address:
          type: string
        city:
          type: string
        province:
          type: string
        country:
          type: string
        phone:
          type: string
        mobile:
          type: string
      required: [name]
...
    StaffMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          $ref: "#/components/schemas/StaffRole"
        active:
          type: boolean
      required: [id, name, role]
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        description:
          type: string
        quantity:
          type: number
        unitPrice:
          type: number
        total:
          type: number
      required: [description, quantity, unitPrice, total]
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNumber:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        clientId:
          type: integer
        client:
          $ref: "#/components/schemas/Client"
        measuredById:
          type: integer
        measuredBy:
          $ref: "#/components/schemas/StaffMember"
        quotedById:
          type: integer
        quotedBy:
          $ref: "#/components/schemas/StaffMember"
        budgetDate:
          type: string
          format: date
        deliveryDate:
          type: string
          format: date
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        downPayment:
          type: number
        notes:
          type: string
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        total:
          type: number
      required: [id, invoiceNumber, status, total]
    CreateInvoiceRequest:
      type: object
      properties:
        clientId:
          type: integer
        measuredById:
          type: integer
        quotedById:
          type: integer
        budgetDate:
          type: string
          format: date
        deliveryDate:
          type: string
          format: date
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        downPayment:
          type: number
        notes:
          type: string
        status:
          $ref: "#/components/schemas/InvoiceStatus"
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              description:
                type: string
              quantity:
                type: number
              unitPrice:
                type: number
              total:
                type: number
            required: [description, quantity, unitPrice, total]
      required: [items]
    PaginatedInvoices:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
    PaginatedClients:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Client"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [data, total, page, limit, totalPages]
...