# Invoice-System-Pro/lib/api-zod/src/generated/types/client.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandro Industrial Invoice System API
 * OpenAPI spec version: 0.1.0
 */

export interface Client {
  id: number;
  code?: string;
  name: string;
  address?: string;
  city?: string;
  province?: string;
  country?: string;
  phone?: string;
  mobile?: string;
  createdAt?: Date;
}