# Invoice-System-Pro/lib/api-zod/src/generated/types/paginatedInvoices.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sandro Industrial Invoice System API
 * OpenAPI spec version: 0.1.0
 */
import type { Invoice } from "./invoice";

export interface PaginatedInvoices {
  data: Invoice[];
  total: number;
  page: number;
  limit: number;
  totalPages: number;
}